moved snakeyaml to another module
