+++ libraries-data-io/pom.xml
@@ -88,7 +88,7 @@
     </dependencies>
 
     <properties>
-        <snakeyaml.version>2.2</snakeyaml.version>
+        <snakeyaml.version>1.30</snakeyaml.version>
         <opencsv.version>5.8</opencsv.version>
         <google-api.version>1.23.0</google-api.version>
         <google-sheets.version>v4-rev493-1.21.0</google-sheets.version>
