+++ libraries-data-io/src/main/java/com/baeldung/libraries/snakeyaml/SafeConstructorLoader.java
@@ -1,12 +1,12 @@
-package com.baeldung.boot;
+package com.baeldung.libraries.snakeyaml;
 
 import org.yaml.snakeyaml.Yaml;
 import org.yaml.snakeyaml.constructor.Constructor;
 
 public class SafeConstructorLoader {
 
-    public MyCustomClass parseYaml(String yamlInput) {
-        Yaml yaml = new Yaml(new Constructor(MyCustomClass.class));
+    public Customer parseYaml(String yamlInput) {
+        Yaml yaml = new Yaml(new Constructor(Customer.class));
         return yaml.load(yamlInput);
     }
