+++ libraries-data-io/src/main/java/com/baeldung/libraries/snakeyaml/VulnerableYamlLoader.java
@@ -1,9 +1,7 @@
-package com.baeldung.boot;
+package com.baeldung.libraries.snakeyaml;
 
-import org.springframework.stereotype.Service;
 import org.yaml.snakeyaml.Yaml;
 
-@Service
 public class VulnerableYamlLoader {
     public Object loadYaml(String yamlContent) {
         Yaml yaml = new Yaml();  // Default constructor - unsafe deserialization
