--- spring-boot-modules/spring-boot-snakeyaml/README.md
@@ -1,8 +1,0 @@
-## Spring Boot Snakeyaml
-
-This module contains articles about Spring Boot with Snakeyaml
-
-## Relevant Articles:
-
-- [Resolving CVE-2022-1471  with the SnakeYAML 2.0 Release](https://drafts.baeldung.com/?p=215658&preview_id=215658&preview_nonce=5b5ad69827&_thumbnail_id=-1&preview=true)
-
