--- spring-boot-modules/spring-boot-snakeyaml/src/main/java/com/baeldung/boot/MyCustomClass.java
@@ -1,22 +1,0 @@
-package com.baeldung.boot;
-
-public class MyCustomClass {
-    private String name;
-    private int age;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-}
