--- spring-boot-modules/spring-boot-snakeyaml/src/main/resources/sample.yaml
@@ -1,2 +1,0 @@
-key: value
-anotherKey: !!java.util.Date {}
